requirements.txt
requests>=2.12.4            # api tests
selenium>=2.53.6            # web tests
Appium-Python-Client>=0.24  # mobile tests
PyYAML==3.11                # page object specification files
six>=1.10.0
screeninfo
